Refactoring changes to architecture and fixed most of the tests.
Improved overall readability by cleaning up code, rearringing elements and variables.
Fixed all the tests related to non geometry processing, so that the development of the targeting system can continue.

Cleaned up factory and code related to creation of geometric vision.
Most of the changes are made so that the current system works with the entity system